requirements.txt
kafka-python==2.0.2
requests==2.26.0
pravega==0.3.1
pyspark==3.1.1
